new unit conventions, examples and refactoring

+++ crates/dslab-compute/src/multicore.rs
@@ -77,7 +77,7 @@ impl RunningComputation {
 
 #[derive(Serialize)]
 pub struct CompRequest {
-    pub flops: u64,
+    pub flops: f64,
     pub memory: u64,
     pub min_cores: u32,
     pub max_cores: u32,
@@ -139,7 +139,7 @@ pub struct DeallocationFailed {
 // ACTORS //////////////////////////////////////////////////////////////////////////////////////////
 
 pub struct Compute {
-    speed: u64,
+    speed: f64,
     cores_total: u32,
     cores_available: u32,
     memory_total: u64,
@@ -150,7 +150,7 @@ pub struct Compute {
 }
 
 impl Compute {
-    pub fn new(speed: u64, cores: u32, memory: u64, ctx: SimulationContext) -> Self {
+    pub fn new(speed: f64, cores: u32, memory: u64, ctx: SimulationContext) -> Self {
         Self {
             speed,
             cores_total: cores,
@@ -163,7 +163,7 @@ impl Compute {
         }
     }
 
-    pub fn speed(&self) -> u64 {
+    pub fn speed(&self) -> f64 {
         self.speed
     }
 
@@ -185,7 +185,7 @@ impl Compute {
 
     pub fn run(
         &mut self,
-        flops: u64,
+        flops: f64,
         memory: u64,
         min_cores: u32,
         max_cores: u32,
@@ -250,7 +250,7 @@ impl EventHandler for Compute {
 
                     let speedup = cores_dependency.speedup(cores);
 
-                    let compute_time = flops as f64 / self.speed as f64 / speedup;
+                    let compute_time = flops / self.speed / speedup;
                     self.ctx.emit_self(CompFinished { id: event.id }, compute_time);
                     self.computations
                         .insert(event.id, RunningComputation::new(cores, memory, requester));
